[feature/issue-702] Code Review
