+++ internal/armada/repository/job.go
@@ -1003,105 +1003,3 @@ if currentClusterId == clusterId then
 end
 return 0
 `)
-
-type reportEvents func([]*api.EventMessage) error
-type AddJobs func([]*api.Job) (SubmitJobResults, error)
-
-func (add AddJobs) ReportQueued(report reportEvents) AddJobs {
-	return func(jobs []*api.Job) (SubmitJobResults, error) {
-		results, err := add(jobs)
-		if err != nil {
-			return nil, err
-		}
-
-		filter := func(result *SubmitJobResult) bool {
-			return result.Error == nil && !result.DuplicateDetected
-		}
-
-		now := time.Now()
-		events := make([]*api.EventMessage, 0, len(jobs))
-		for _, result := range results.Filter(filter) {
-			events = append(events, &api.EventMessage{
-				Events: &api.EventMessage_Queued{
-					Queued: &api.JobQueuedEvent{
-						JobId:    result.JobId,
-						Queue:    result.SubmittedJob.Queue,
-						JobSetId: result.SubmittedJob.JobSetId,
-						Created:  now,
-					},
-				},
-			})
-		}
-
-		if err := report(events); err != nil {
-			return nil, fmt.Errorf("failed to report queued jobs")
-		}
-
-		return results, nil
-	}
-}
-
-func (add AddJobs) ReportDuplicate(report reportEvents) AddJobs {
-	return func(jobs []*api.Job) (SubmitJobResults, error) {
-		results, err := add(jobs)
-		if err != nil {
-			return nil, err
-		}
-
-		filter := func(result *SubmitJobResult) bool {
-			return result.Error == nil && result.DuplicateDetected
-		}
-		now := time.Now()
-		events := make([]*api.EventMessage, 0, len(jobs))
-		for _, result := range results.Filter(filter) {
-			events = append(events, &api.EventMessage{
-				Events: &api.EventMessage_DuplicateFound{
-					DuplicateFound: &api.JobDuplicateFoundEvent{
-						JobId:         result.JobId,
-						Queue:         result.SubmittedJob.Queue,
-						JobSetId:      result.SubmittedJob.JobSetId,
-						Created:       now,
-						OriginalJobId: result.JobId,
-					},
-				},
-			})
-		}
-
-		if err := report(events); err != nil {
-			return nil, fmt.Errorf("failed to report duplicate jobs")
-		}
-
-		return results, nil
-	}
-}
-
-func (add AddJobs) ReportSubmitted(report reportEvents) AddJobs {
-	return func(jobs []*api.Job) (SubmitJobResults, error) {
-		results, err := add(jobs)
-		if err != nil {
-			return nil, err
-		}
-
-		now := time.Now()
-		events := make([]*api.EventMessage, 0, len(jobs))
-		for _, result := range results {
-			events = append(events, &api.EventMessage{
-				Events: &api.EventMessage_Submitted{
-					Submitted: &api.JobSubmittedEvent{
-						JobId:    result.JobId,
-						Queue:    result.SubmittedJob.Queue,
-						JobSetId: result.SubmittedJob.JobSetId,
-						Created:  now,
-						Job:      *result.SubmittedJob,
-					},
-				},
-			})
-		}
-
-		if err := report(events); err != nil {
-			return nil, fmt.Errorf("failed to report queued jobs")
-		}
-
-		return results, nil
-	}
-}
