+++ internal/armada/server/jobs_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 	"testing/quick"
 
-	"github.com/G-Research/armada/internal/common/auth/authorization"
 	"github.com/G-Research/armada/pkg/api"
 )
 
@@ -43,7 +42,7 @@ func TestNewJobs(t *testing.T) {
 
 			return true
 		},
-		"erorr": func(req api.JobSubmitRequest, principal string) bool {
+		"error": func(req api.JobSubmitRequest, principal string) bool {
 			new := NewJobs(
 				principal,
 				func(c context.Context, queueName string) ([]string, error) {
@@ -121,55 +120,6 @@ func TestNewJobsValidate(t *testing.T) {
 				return false
 			}
 
-			return true
-		},
-	}
-
-	for name, property := range properties {
-		t.Run(name, func(tb *testing.T) {
-			if err := quick.Check(property, nil); err != nil {
-				tb.Fatal(err)
-			}
-		})
-	}
-}
-
-func TestGetQueueOwnership(t *testing.T) {
-	properties := map[string]interface{}{
-		"success": func(qname string, qeueu api.Queue, groups []string) bool {
-			getOwnership := GetQueueOwnership(
-				func(queueName string) (*api.Queue, error) {
-					if queueName != qname {
-						return nil, fmt.Errorf("invalid queue name")
-					}
-					return &qeueu, nil
-				},
-				func(c context.Context, o authorization.Owned) (bool, []string) {
-					return true, groups
-				},
-			)
-
-			ownerships, err := getOwnership(context.Background(), qname)
-			if err != nil {
-				t.Errorf("failed to retrieve queue ownership: %s", err)
-				return false
-			}
-			return reflect.DeepEqual(ownerships, groups)
-		},
-		"error": func(qname string) bool {
-			getOwnership := GetQueueOwnership(
-				func(queueName string) (*api.Queue, error) {
-					return nil, fmt.Errorf("")
-				},
-				func(c context.Context, o authorization.Owned) (bool, []string) {
-					return true, nil
-				},
-			)
-
-			if _, err := getOwnership(context.Background(), qname); err == nil {
-				t.Errorf("failed to handle an error")
-				return false
-			}
 			return true
 		},
 	}
